Skip tests if `LOCAL_S3` env var is not set
Skip the test suit when the local S3 flag is not set. This is because,
otherwise the tests will certainly fail as the required config will be
empty. Further work is needed to make tests run both on local and remote
S3. For now skip if local S3 flag is absent.

Relates to:
 - https://github.com/ipfs/go-ds-s3/pull/171
 - https://github.com/orgs/ipfs/projects/12#card-58209321